script/requirements.txt
pyserial
numpy
keyboard
scipy
